requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.31.0
google-auth-oauthlib==1.2.2
